Don't show bopomofo state during composition
[!] This is a commit preparing for oncoming braille unicode display during
    composition.

The state of bopomofo is now maintained in bpmf_cumulative_str, a buffer
defined by braille chewing. The feedback from append_brl is performed on
the buffer, and the buffer pushes its content to chewing when the internal
braille state is reset.

Note that the behavior of Esc is overwritten when bpmf_cumulative_str is
not empty to keep consistency with previous versions.