Let the user select the style of internal state display
There are two styles of displaying the internal state, bopomofo and braille
unicode. 145+space is assigned as the hot key to switch between them.

One extra field display_ucbrl is added into class brl_buf_state. It is
always interpreted as a boolean value. Its name does not start with an
underscore, meaning that modification outside the class is accepted.

* It is thanks to "·£¯S¨µ°j ¶À°¶»¨" <hurt.nzsmr@gmail.com> for the
  proposal of 145+space.