Let 145+space have three states
An additional state is that nothing will be displayed to represent the
state of composition. This is used to input nomally when an NVDA user uses
Braille Chewing in the latest Skype.

In fact, this problem may be due to either NVDA or Skype. However, it seems
that the problem will not be solved soon.